feat(test): step2 - separate java pakage
step2 - separate java pakage
